// src/lib/types.ts
export interface ShapeFabricRefs {
  rect: Rect
  label: FabricText
  dims: FabricText
}

export interface LineFabricRefs {
  line: Line
  label: FabricText
}

export interface MaskFabricRefs {
  rect: Rect
}

export interface ImageFabricRefs {
  image: FabricImage
}

export type FabricRefs =
  | ShapeFabricRefs
  | LineFabricRefs
  | MaskFabricRefs
  | ImageFabricRefs

// ============================================
// Serialization (JSON export/import)
// ============================================
interface SerializedBase {
  id: number
  type: ObjectType
  name: string
  left: number
  top: number
  scaleX: number
  scaleY: number
  angle: number
}

export interface SerializedShape extends SerializedBase {
  type: 'shape'
  widthM: number
  heightM: number
  color: string
  baseWidthPx: number
  baseHeightPx: number
  width: number
  height: number
}

export interface SerializedLine extends SerializedBase {
  type: 'line'
  x1: number
  y1: number
  x2: number
  y2: number
  lengthM: number
  color: string
  strokeWidth: number
}

export interface SerializedMask extends SerializedBase {
  type: 'mask'
  width: number
  height: number
}

export interface SerializedImage extends SerializedBase {
  type: 'backgroundImage' | 'overlayImage'
  imageData: string
  originX: string
  originY: string
}

export type SerializedObject =
  | SerializedShape
  | SerializedLine
  | SerializedMask
  | SerializedImage